Add final principles (cryptograpically signed by >50% of CCs)
Previous commits:

First draft of principles

Updated based on feedback so far

Medium:
- Use more active words instead if "aim" and "will"
- Remove "also known as web3."
- Replace "Ethereum is to be censorship resistant, and we should provide the
same guarantees across Status." with ""we abide by the cryptoeconomic design principle of censorship resistance""

Minor:
- Remove linebreak

Bring back "aim to maximize ...freedoms", otherwise it reads weird

Medium changes:
- Use Roman numerals to ease future translations (referencing principles)
- Free culture => Free use
- Privacy: change section to emphasize pseudo anonymity:
"For us, it's essential to protect privacy in both communications and
transactions, as well as being a Pseudo-Anonymous platform. Additionally, we
strive to provide the right of total anonymity."
- Continuance: Add incentivized to "We create software incentivized to continue to exist and improve, without the stewardship of a single entity or any of the current team members."

Add interoperability to Principles/Inclusivity

enabling its users to communicate freely with each other no matter the software they using.
For Status this would mean that we would try to make accessible the "import" and "export"  of user's content from other "Transactor softwares", such as Status to other such as Mist, that would also implement this interoperability capabilities.
As we are open source this is easy as agree in use the same methods, but we need to achieve some coordination with other wallet providers to provide at end the best user experience, and user are freely to choice which software they prefer, and easily move on to other if they want.
Interoperability should be achievable in Status through use of ERC725 (Identity) and other future account interfaces.

Medium changes:

- Adjust free culture / free use header to Openness

- Adjust body section on resourcefulness:

X. Resourcefulness
We are relentlessly resourceful. As we grow and have ready access to capital, it is our obligation to token holders to fight bureaucracy and inefficiencies. This means solving problems in the most effective way possible at lower economic costs (in terms of capital, time and resources).

- Liberty: Add coercion-resistant sentence
...and economic freedoms. This includes being coercion-resistant.

- Add back agnostic sentence as it seems more related here than under Liberty, and has come up a few times in conversations as something these principles should capture.
II. Censorship resistance
We enable free flow of information. No content is under surveillance. We abide by the cryptoeconomic design principle of censorship resistance. Even stronger, Status is an agnostic platform for information.

Minor capital

Minor clarifications on scope

- V: We strive for complete openness and symmetry of information _within the organization_.

- X: As we grow and have ready access to capital, it is our obligation to token holders to fight bureaucracy and inefficiencies _within the organization_.